config page design done

+++ LibraryApp.Client/src/app/config/books/books.component.ts
@@ -5,13 +5,14 @@ import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { BookModalComponent } from './book-modal.component';
 import { ConfigService } from '../config.service';
 import { IBook } from '../../shared/models/book';
+import { WarningDialogComponent } from 'src/app/shared/dialogs/warning-dialog.component';
 
 @Component({
     templateUrl: './books.component.html',
 })
 export class BooksComponent implements OnInit {
     public books: IBook[];
-    
+
     constructor(
         private modalService: NgbModal,
         private configService: ConfigService,
@@ -23,20 +24,27 @@ export class BooksComponent implements OnInit {
     }
 
     public deleteBook(book: IBook) {
-        this.configService.deleteBook(book.id)
-            .subscribe(result => {
-                if (result) {
-                    this.loadBooks();
-                }
-            })
+        const modalRef = this.modalService.open(WarningDialogComponent, { size: 'sm' });
+
+        modalRef.result.then((result) => {
+            if (result == true) {
+                this.configService.deleteBook(book.id)
+                    .subscribe(result => {
+                        if (result) {
+                            this.loadBooks();
+                        }
+                    });
+            }
+        });
     }
 
-    public updateBook(book: IBook) {
-        this.openBookModal(book.id);
+    public updateBook(user: IBook) {
+        this.openBookModal(user.id);
     }
 
     public addBook() {
         this.openBookModal(undefined);
+
     }
 
     public loadBooks() {
@@ -58,10 +66,10 @@ export class BooksComponent implements OnInit {
     }
 
     public getBookImage(imageString: any) {
-        if(!!imageString){
-            return this.sanitizer.bypassSecurityTrustResourceUrl('data:image/jpg;base64,' 
-            + imageString);
+        if (imageString) {
+            return imageString;
         }
+        return "./assets/images/no-avatar.png"
     }
 
 }
