+++ LibraryApp.Client/src/app/config/users/users.component.ts
@@ -4,6 +4,7 @@ import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { IUser } from '../../shared/models/user';
 import { ConfigService } from '../config.service';
 import { UserModalComponent } from './user-modal.component';
+import { WarningDialogComponent } from '../../shared/dialogs/warning-dialog.component';
 
 @Component({
     templateUrl: './users.component.html',
@@ -20,12 +21,18 @@ export class UsersComponent implements OnInit {
          this.loadUsers();
     }
     public deleteUser(user: IUser) {
-        this.configService.deleteUser(user.id)
-        .subscribe(result => {
-            if (result) {
-                this.loadUsers();
-            }
-        })
+        const modalRef = this.modalService.open(WarningDialogComponent, { size: 'sm' });
+        
+                modalRef.result.then((result) => {
+                    if (result == true) {
+                        this.configService.deleteUser(user.id)
+                        .subscribe(result => {
+                            if (result) {
+                                this.loadUsers();
+                            }
+                        });
+                    }
+                });
     }
 
     public updateUser(user: IUser) {
@@ -56,9 +63,10 @@ export class UsersComponent implements OnInit {
     } 
 
     public getUserImage(imageString: any) {
-        if(!!imageString){
-            return this.sanitizer.bypassSecurityTrustResourceUrl('data:image/jpg;base64,' 
-            + imageString);
+        if(imageString){
+            return imageString;
         }
+        return "./assets/images/no-avatar.png"
+        
     }
 }
